requirements.txt
git+https://github.com/Rapptz/discord.py
discord.py
PyNaCl==1.3.0
dnspython
async-timeout
boto